requirements.txt
# ── Core framework ───────────────────────────────────────────────────────────
fastapi>=0.111
uvicorn[standard]>=0.29
websockets>=12.0
pydantic>=2.7
python-multipart>=0.0.9

# ── LLM & AI ────────────────────────────────────────────────────────────────
openai>=1.30
anthropic>=0.39

# ── Search & data sources ───────────────────────────────────────────────────
findpapers @ git+https://github.com/jonatasgrosman/findpapers.git
requests>=2.31
pyalex>=0.14
scholarly>=1.7

# ── Async & networking ──────────────────────────────────────────────────────
aiohttp>=3.9
aiohttp-socks>=0.8
brotli>=1.1
brotlicffi>=1.0
aiolimiter>=1.1
tenacity>=8.2

# ── PDF & text processing ───────────────────────────────────────────────────
pypdf>=4.0
pymupdf>=1.24
readability-lxml>=0.8
beautifulsoup4>=4.12
lxml>=5.0

# ── Data & export ───────────────────────────────────────────────────────────
pandas>=2.0
openpyxl>=3.1

# ── Integrations ────────────────────────────────────────────────────────────
notion-client>=3.0

# ── Config & utilities ──────────────────────────────────────────────────────
python-dotenv>=1.0
tqdm>=4.66

# ── Legacy Streamlit UI (optional) ──────────────────────────────────────────
streamlit>=1.30

# ── Testing ─────────────────────────────────────────────────────────────────
pytest>=8.0
